+++ __tests__/ImportOnlineActions.test.js
@@ -1,0 +1,13 @@
+import configureMockStore from 'redux-mock-store';
+import thunk from 'redux-thunk';
+import consts from '../src/js/actions/ActionTypes';
+import * as importOnline  from '../src/js/actions/ImportOnlineSearchActions';
+
+const middlewares = [thunk];
+const mockStore = configureMockStore(middlewares);
+
+describe('ConfirmOnlineActions.confirmOnlineAction', () => {
+    test('loadOnline.openManifest should return an error if no link is specified', () => {
+      assert(importOnline.confirmOnlineAction).not.isNotNull();
+    });
+});
